project payroll almost done

+++ projectroot/src/main/java/com/cebedo/pmsys/bean/MultipartBean.java
@@ -4,40 +4,40 @@ import org.springframework.web.multipart.MultipartFile;
 
 public class MultipartBean {
 
-	private MultipartFile file;
-	private long projectID;
-	private String description;
+    private MultipartFile file;
+    private long projectID;
+    private String description;
 
-	public MultipartBean() {
-		;
-	}
+    public MultipartBean() {
+	;
+    }
 
-	public MultipartBean(int id) {
-		setProjectID(id);
-	}
+    public MultipartBean(long id) {
+	setProjectID(id);
+    }
 
-	public String getDescription() {
-		return description;
-	}
+    public String getDescription() {
+	return description;
+    }
 
-	public void setDescription(String description) {
-		this.description = description;
-	}
+    public void setDescription(String description) {
+	this.description = description;
+    }
 
-	public MultipartFile getFile() {
-		return file;
-	}
+    public MultipartFile getFile() {
+	return file;
+    }
 
-	public void setFile(MultipartFile file) {
-		this.file = file;
-	}
+    public void setFile(MultipartFile file) {
+	this.file = file;
+    }
 
-	public long getProjectID() {
-		return projectID;
-	}
+    public long getProjectID() {
+	return projectID;
+    }
 
-	public void setProjectID(long projectID) {
-		this.projectID = projectID;
-	}
+    public void setProjectID(long projectID) {
+	this.projectID = projectID;
+    }
 
 }
